ai_meeting_note_extractor_process/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.3",
    "openai>=2.31.0",
    "openai-agents>=0.14.0",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.44.0",
    "tenacity>=8.5.0",
]
